[refactor] combined Heading components
