+++ components/HeaderBar.js
@@ -5,7 +5,7 @@ export default function HeaderBar({ children }) {
   return (
     <div className="sticky top-0 left-0 z-30 flex min-h-fit w-screen items-center justify-center gap-4 bg-white/80 backdrop-blur dark:bg-black/60">
       <Link href={"/"} passHref>
-        <div className={`flex cursor-pointer gap-4 py-4`}>
+        <div className={`flex cursor-pointer items-center gap-4 py-4`}>
           <LogoIcon size={"sm"} />
           {children}
         </div>
