+++ components/Heading/Heading.js
@@ -1,0 +1,46 @@
+import { useRouter } from "next/router";
+
+export default function Heading({ size }) {
+  const router = useRouter();
+  const currentPath = router.pathname;
+  const heading = () => {
+    if (currentPath === "/strains") {
+      return {
+        viewBox: "0 0 131 52",
+        svg: (
+          <path d="M5.6 51.4c-1.6 0-3-.3-4-.8-.7-.4-1-.9-1-1.4 0-.5 0-.9.3-1.2.3-.3.7-.5 1.1-.5.5 0 1 .2 1.9.5.9.4 1.4.5 1.7.5a3 3 0 0 0 2.1-1c.7-.7 1.2-1.7 1.6-2.8a17.2 17.2 0 0 0 .3-11.5c-.6-2-1.3-4-2.3-6l-1-2-2.5-5c-.6-1.3-1-2.7-1.5-4.1-.7-2-1-4-1-6 0-1.5.3-3 .8-4.3a9 9 0 0 1 2.4-3.5c1-1 2.4-1.5 4-1.5a4 4 0 0 1 2.2.7c.8.4 1.3 1 1.6 1.7l-1.1 1.3c-1.1 0-2-.2-2.5-.6-.8 0-1.5.4-2 1.1A6.8 6.8 0 0 0 5 9.7a20.5 20.5 0 0 0 2.4 9.7l4.5 9a24.6 24.6 0 0 1 2.4 11c0 2.7-.7 5.4-2 7.8a8.3 8.3 0 0 1-3 3 6.8 6.8 0 0 1-3.7 1.2ZM25.2 49v-2.5L25 44l.1-2.1-1-14c-.5-9-.8-14.6-.8-17.1-2.5-.1-4.1-.4-4.8-.8-.4-.3-.7-.7-.7-1.2 0-.3.2-.6.7-.8.4-.3.9-.5 1.4-.5 1.2 0 2.6 0 4.3.3l4.4.3h1c1.3 0 2 .5 2 1.5 0 .5-.2 1-.7 1.3-.4.3-.9.4-1.4.4l-1.7-.1-1.3-.3v.3l-.1.3c0 3.8.4 9.9 1 18.2.8 7.8 1.1 13.9 1.1 18.2 0 .5-.2 1-.7 1.3-.4.4-1 .6-1.5.6l-1.1-.8Zm24 2c-1 0-1.8-1-2.6-3a35 35 0 0 1-2-6.8l-.4-2.2c-1.3-6.1-2-9.6-2.4-10.4H41v6.9c0 5.7.4 10.1 1.2 13.3l-.7 1.8h-1.8l-.8-1v-4.3c0-3.8-.2-9.6-.5-17.4-.4-8.4-.6-14.2-.6-17.5 0-2 .7-3 2.2-3 2.3 0 4.1.5 5.6 1.7a6 6 0 0 1 2.2 5v4.5c0 .6 0 1.4-.3 2.2-.2.8-.4 1.5-.7 2l-1.2 2.3-1 2 .3 1.5.9 4.5c.6 3.3 1.1 6 1.7 8 .6 2 1.2 4 2 5.6l.4.7c.5.9.7 1.6.7 2 0 .5-.2.9-.4 1.2-.3.3-.6.5-1 .5Zm-8-25.6c1-.4 1.8-1.7 2.5-3.8a22 22 0 0 0 1-5.9 14 14 0 0 0-.5-3.8 3 3 0 0 0-1.1-1.6c-.5-.4-1-.6-1.8-.6l-.1.4-.2.5.2 7.6v7.2Zm28.2 25.2c-1 0-1.7-.2-2.1-.5-.5-.3-.7-.8-.7-1.6a55 55 0 0 0-.7-6.6l-.2-1.3-6.4-1-1 6c-.3 2-.7 3.5-1.3 4.1-.3.4-.7.6-1.2.6-.6 0-1-.1-1.5-.4-.4-.4-.6-.8-.6-1.2l.1-.4.1-.2a284.7 284.7 0 0 0 7-38.9c.1-1 .4-1.6.7-2h2.2c.6 0 1.1 0 1.6.4.4.2.6.6.6 1.1 0 3.3.4 8 1.3 14.5.9 6.6 1.3 11.2 1.3 13.8l.3 2 .4 2 .6 3.8c.4 2 .6 3.3.6 4 0 1-.4 1.5-1 1.8Zm-4.2-12.3-.2-2.4c0-3.8-.3-8.7-1-14.7l-.6-6.7c-1.9 10-3.1 17.6-3.8 22.8.6.4 2.5.7 5.6 1Zm14 11.6c-3.7 0-5.6-.6-5.6-1.9 0-1 .7-1.5 2-1.5h1.3c.4 0 .6 0 .7-.3.2-.2.2-.6.2-1.1a68 68 0 0 0-.6-9 274.2 274.2 0 0 1-.5-18l.1-3.2c0-1.4 0-2.6-.2-3.6a7 7 0 0 1-1.5.3c-1 0-1.5-.5-1.5-1.5 0-.5.3-1 1-1.4A7 7 0 0 1 77 8c1.7-.4 3.2-.5 4.5-.5.5 0 1 .2 1.3.6.4.3.6.7.6 1.2 0 .4-.2.8-.5 1.2-.2.3-.6.5-1 .5h-1.1c-.5 0-.8 0-1 .3-.2.1-.2.5-.2 1a333.4 333.4 0 0 0 1.2 25.4c.2 2 .4 4 .4 6v2.5h1.2c1.8 0 2.7.7 2.7 2.2v.4l-1.6 1-.4.1h-1.3l-2.6.1Zm28.7.3c-.5 0-1-.1-1.2-.4l-.6-1.5-.3-1.2v-.3a352.4 352.4 0 0 0-11.3-32.6v32.2l.3.5.2.4.2.7c0 .5-.2 1-.6 1.4a2 2 0 0 1-1.4.5c-1.3 0-2-1.3-2-3.9V35a854.1 854.1 0 0 1 0-16.9c.2-2.7.2-4.6.2-6 0-2 .8-3.5 2.3-4.7v.4l.5-.1c1 0 2.3 1.7 3.8 5.1a88 88 0 0 1 3.2 8.7c1 3 2.2 7.3 3.8 12.9l.6 2.3L108 44c.8-9.2 1.3-20 1.3-32.5v-.8l-.1-.9c0-1.2.5-1.9 1.5-1.9.6 0 1 .3 1.4.7.3.5.5 1 .5 1.8v4.8a590 590 0 0 1-.8 21.4c-.4 7.7-1 12-1.5 13l-2.3.5Zm14.3.7c-.8-.1-1.4-.4-2-.8-.5-.5-.8-1-.9-1.7l.7-1c.3 0 .7 0 1.2.2l1.5.2c.8 0 1.4-.4 2-1 .7-.7 1.1-1.6 1.4-2.8.6-2 .9-3.8.9-5.5a18 18 0 0 0-2.7-8.9l-.7-1.2-2-3.7a22.3 22.3 0 0 1-2.2-8.6c0-1.9.4-3.9 1.2-6a6 6 0 0 1 1.8-2.3c.8-.6 1.7-1 2.7-1 .8 0 1.5.3 2.1.8.7.5 1 1 1 1.8 0 .3 0 .6-.4.9-.3.2-.6.3-1 .3-.5 0-1-.2-1.4-.4-1 0-1.7.9-2.3 2.5a13 13 0 0 0-.7 3.7 16.3 16.3 0 0 0 2 7.2L126 26l2.2 4.3a20.2 20.2 0 0 1 2 9c0 2.5-.6 5-1.8 7.3a7.3 7.3 0 0 1-2.5 3c-1.1.6-2.3 1-3.6 1v.2Z" />
+        ),
+      };
+    } else {
+      return {
+        viewBox: "0 0 212 93.2",
+        svg: (
+          <path d="M5.7 92Q.2 92 .2 89l.3-14.7.2-14.6-.3-24.2Q0 19.5 0 11.3v-.7a25.6 25.6 0 0 1 .6-5.7 3.6 3.6 0 0 1 3.8-3.2q8.3 0 13.2 5.3t5 14.2a25.4 25.4 0 0 1-3 11.2 23 23 0 0 1-6.7 8.7 28 28 0 0 1 10.2 12 39.2 39.2 0 0 1 3.2 16.2q0 4.1-2.8 10.3Q20 87 14 90.2A17.3 17.3 0 0 1 6.2 92h-.5ZM7 39.3c2-.5 3.8-2.2 5.3-5.1a50.2 50.2 0 0 0 3.3-8.1 31.6 31.6 0 0 1 1.2-3.3v-4.2a11 11 0 0 0-2.6-7c-1.8-2.1-3.8-3.3-6-3.3C7 8.3 6.4 9.6 6.4 12q0 7.9.3 14Zm.5 46.5a13.6 13.6 0 0 0 9.3-5.4A17.3 17.3 0 0 0 20 69.6a52.3 52.3 0 0 0-.9-10 22.1 22.1 0 0 0-2.3-7 26.3 26.3 0 0 0-2.6-3.9 8 8 0 0 0-2.3-2l-1-.4H6.8q.6 13.7.6 39.5Zm50.9 7.4a13.4 13.4 0 0 1-11.7-6.5 41.5 41.5 0 0 1-6.6-16A138.4 138.4 0 0 1 37.3 49q-.6-10.5-.6-22.8v-9.9l.1-8.8a3.2 3.2 0 0 1-1-2.4 3.6 3.6 0 0 1 1-2.5 3.2 3.2 0 0 1 2.5-1.2q4.3 0 4.3 8.5 0 5.5.6 20.7.5 15.3.5 20.8a30.2 30.2 0 0 1 1.2 6.2 70.8 70.8 0 0 1 .5 7.6 31.6 31.6 0 0 0 .9 6.5 50 50 0 0 0 2.3 7.8 20.2 20.2 0 0 0 4 6.8A7.2 7.2 0 0 0 59 89a4 4 0 0 0 2.8-1 5.4 5.4 0 0 0 1.5-3.2 31 31 0 0 0 .7-6.5V55a94.6 94.6 0 0 0-.8-11.2q-.7-7.5-.8-11.3v-4.3l-.2-3.4q0-3.4-.4-10.2-.4-6-.4-10.2a2.8 2.8 0 0 1 1.4-2.3 5.2 5.2 0 0 1 3-1l3 2.8V6a15.9 15.9 0 0 1-.5 3.5l-.5 3.8q0 10.8 1.2 29.9 1 20 1 31a41.7 41.7 0 0 1-1 10.7 12.5 12.5 0 0 1-2 4.4 10.5 10.5 0 0 1-3.4 2.8 10.3 10.3 0 0 1-5.2 1.2ZM90.9 92q-5.5 0-5.5-3l.3-14.7.2-14.6-.3-24.2q-.4-16-.4-24.2v-.7a26.2 26.2 0 0 1 .6-5.7 3.6 3.6 0 0 1 3.8-3.2q8.4 0 13.2 5.3t5 14.2a25.4 25.4 0 0 1-3 11.2 23 23 0 0 1-6.7 8.7 27.8 27.8 0 0 1 10.1 12 39 39 0 0 1 3.3 16.1q0 4.2-2.8 10.4-3.5 7.4-9.4 10.6a17.4 17.4 0 0 1-8 1.8H91Zm1.2-52.7c2-.4 3.8-2.2 5.3-5a50.2 50.2 0 0 0 3.3-8.2 31.6 31.6 0 0 1 1.2-3.3v-4.2a11 11 0 0 0-2.6-7c-1.8-2.1-3.8-3.3-6-3.3-1.2 0-1.7 1.3-1.7 3.7q0 8 .2 14Zm.5 46.5a13.6 13.6 0 0 0 9.3-5.4 17.3 17.3 0 0 0 3.4-10.8 52.3 52.3 0 0 0-.9-10 22.1 22.1 0 0 0-2.3-7 26.3 26.3 0 0 0-2.6-3.9 7.9 7.9 0 0 0-2.3-2l-1-.4H92q.6 13.7.6 39.5Zm53.7 5.2a4.5 4.5 0 0 1-2.8-.7 2.9 2.9 0 0 1-.9-2.4V72.3a16.6 16.6 0 0 0-5.8-1.9 64.5 64.5 0 0 0-8.1-1 20.4 20.4 0 0 0-2 6l-.4 2.4-.7 3.7a39.6 39.6 0 0 1-1.3 5 16.8 16.8 0 0 1-1.4 3.3q-4.6 0-5.3-2.4 1.3-2 3-11.2l4-21.8q2.6-15.5 4.8-26.5A151.4 151.4 0 0 1 134 9.6q2.7-8.2 5.2-8.2a6.5 6.5 0 0 1 3.8 1q0 4.7 2.3 23 2.4 17.8 2.4 24.3A177.9 177.9 0 0 0 149 68q1.2 12 1.2 19.1 0 4-4 4Zm-6.2-25.6a10.1 10.1 0 0 0 2.4-.4q-.3-5.2-1.2-13-2.7-26.9-2.7-33v-1q-1.2 1.3-3.3 10.9t-3.9 20.5a145.7 145.7 0 0 0-1.9 14.6 38.6 38.6 0 0 0 10.6 1.4Zm33.3 26.4a3.1 3.1 0 0 1-2.8-1.7 8.3 8.3 0 0 1-1-4.2q0-4 .5-9.7l.2-3 .2-6.4q0-4.2-.8-18.6-.8-14.4-.8-18.6v-4.5l.2-5.8q0-8.7-1.2-11.1a27.4 27.4 0 0 0-3 .3l-3.4.3a4.4 4.4 0 0 1-2.8-1 2.8 2.8 0 0 1-1.2-2.2c0-1 1-1.8 2.9-2.5q4.2-1.5 14.2-2l4.9-.3a4 4 0 0 1 2.6 1 3 3 0 0 1 1.2 2.4 3 3 0 0 1-1 2.3 3.7 3.7 0 0 1-2.6 1l-2.1-.2a17.5 17.5 0 0 0-2.2-.1h-1.2a145.9 145.9 0 0 1 .8 14.6l-.1 7.6-.1 7.8q0 5.7.4 13.7.6 9.9.6 13.7 0 4.5.5 10.7.5 6.3.5 10.7 0 5.9-3.5 5.9Zm21.4.3q-2.1 0-3.4-2.9 2.6-4 6-21.6t5.8-36.4q2.3-18.5 2.6-24.6l-4.3.4a19.3 19.3 0 0 1-2.5.1q-6.4 0-6.4-3.7V3l1.4-2V.8a10.9 10.9 0 0 0 3.5.5 32 32 0 0 0 5.6-.7 50.6 50.6 0 0 1 5.6-.6 3 3 0 0 1 2.3 1 3.6 3.6 0 0 1 .9 2.5q0 31.3-13.2 82.5l2.6-.1a24.4 24.4 0 0 1 3.7-.3q6.3 0 6.3 3.6v.6l-2 1.8-7.2.2q-3.3.3-7.4.3Z" />
+        ),
+      };
+    }
+  };
+
+  let height;
+  switch (size) {
+    case "sm":
+      height = "h-16";
+      break;
+    case "md":
+      height = "h-24";
+      break;
+    case "lg":
+      height = "h-32";
+      break;
+  }
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      preserveAspectRatio="xMinYMin"
+      viewBox={heading().viewBox}
+      className={`${size ? height : "h-16"} dark:fill-white`}
+    >
+      {heading().svg}
+    </svg>
+  );
+}
